feat: build image from alpine
User part of `caddyserver/caddy-docker` Dockerfile to build a docker image.
Instead of copy the binary `caddy` build with modules to the `caddyserver/caddy-docker` image.
Because of overlayfs, the image contain part of previous image, if I just copy the new binary to it, it will grown bigger.
I want this smaller for using the upx to compress binary.